pyproject.toml
[project]
name = "unike"
version = "3.1.0"
description = "基于 OpenKE-PyTorch 开发的知识图谱嵌入工具包"
readme = "README.md"
authors = [
    { name = "LuYF-Lemon-love", email = "3555028709@qq.com" },
    { name = "wangtao", email = "wangtao.cpu@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.20.3",
    "packaging>=24.0",
    "pandas>=1.1.5",
    "tqdm>=4.67.1",
    "wandb>=0.18.7",
    "torchdata==0.7.0",
    "numpy==1.24.0",
    "dgl==2.4.0+cu121",
]

[project.scripts]
unike = "unike:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = ["docs", "examples", "benchmarks", ".github"]
